Allow resetting cursor style to terminal default
This adds `cursor::DefaultStyle`, which will reset the cursor style back to the terminal default. The default is normally a block cursor, which is why this might seem to give the same results as `cursor::BlinkingBlock`. However, if you change the cursor style in the terminal settings (I tested with Gnome Terminal), then you will see an effect of this escape sequence.

This [Super User answer](https://superuser.com/a/1302503/22077) links to more documentation and discussion around this feature.